+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
  - `asley.machina_extensions.forum` application to extend django-machina's
    forum model.
  - automatic forum creation based on the context of the LTI launch request
+ - Instructor and Administrator LTI roles have special permissions on their
+   forum.
 
 ### Changed